requirements.txt
python-docx>=1.1.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
